requirements.txt
Flask==2.3.3
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
ftfy
sendgrid==6.11.0